requirements.txt
ollama
google-generativeai
python-dotenv
